Improve fast ghosts

+++ src/games/shoot/game_text.c
@@ -327,7 +327,7 @@ void handleLevelBonus(uint16_t bonus)
     {
         _XL_PRINTD(XSize/2-1,YSize/2,3,i);
         points+=5;
-        SHORT_SLEEP(1+(i>>6));
+        SHORT_SLEEP((uint8_t) (1+(i>>6)));
         displayScoreStats();
         _XL_TICK_SOUND();
     }
